JavaSVN 1.0.4 release

+++ javasvn/src/org/tmatesoft/svn/core/io/diff/SVNDiffWindowApplyBaton.java
@@ -36,8 +36,8 @@ public class SVNDiffWindowApplyBaton {
     OutputStream myTargetStream;
 
     long mySourceViewOffset;
-    int mySourceViewLength;
-    int myTargetViewSize;
+    long mySourceViewLength;
+    long myTargetViewSize;
 
     byte[] mySourceBuffer;
     byte[] myTargetBuffer;
